refactor(authentication): move threshold checks to utils refactor(authentication): rename utils file
